\name{describe.oprobit}
\alias{describe.oprobit}
\title{describe a ologit model to zelig...}
\usage{describe.oprobit(...)}
\description{describe a \code{ologit} model to zelig}
\value{a list to be processed by \code{as.description}}
\author{matt owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
