\name{ordinal.zelig-package}
\title{Ordinal Binomial Models in Zelig...}
\description{Ordinal Binomial Models in Zelig}
\details{\tabular{ll}{
Package: \tab ordina.zelig\cr
Version: \tab 1.0-1\cr
Date: \tab 2011-05-04\cr
Depends: \tab Zelig, MASS\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://gking.harvard.edu/zelig\cr
}

Add-on pack for Zelig, containing ordinal models:
\tabular{ll}{
ologit: \tab Ordinal Logit Regression \cr
oprobit: \tab Ordinal Probit Regression \cr
}}
\alias{ordinal.zelig}
\docType{package}
\author{Matt Owen \email{mowen@iq.harvard.edu}, Kosuke Imai, Olivia Lau,
and Gary King 

Maintainer: Matt Owen \email{mowen@iq.harvard.edu}}
\keyword{package}
