\name{param.oprobit}
\alias{param.oprobit}
\title{Param Method for the oprobit Zelig Model...}
\usage{\method{param}{oprobit}(obj, num=1000, ...)}
\description{Param Method for the \code{oprobit} Zelig Model}
\note{This method is used by the \code{oprobit} Zelig model}
\value{a list to be cast as a \code{parameters} object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a \code{zelig} object}
\item{num}{an integer specifying the number of simulations to sample}
\item{...}{ignored parameters}}
