\name{plot.sim.ologit}
\alias{plot.sim.ologit}
\title{plot...}
\usage{plot.sim.ologit(x, ..., alt.color)}
\description{plot.sim.ologit}
\value{function is used for its side-effects}
\arguments{\item{x}{a \code{sim.ologit} object}
\item{...}{ignored parameters}
\item{alt.color}{list of colors to print with}}
