\name{zelig.ordinal-package}
\alias{zelig.ordinal-package}
\alias{zelig.ordinal}
\docType{package}
\title{
Implements Ordinal Regression Models in Zelig
}
\description{
  Provides Ordinal Logit and Probit Regression Models in Zelig, along with
  a set of plotting, quantity of interest, and summary functions.
}
\details{
\tabular{ll}{
Package: \tab zelig.ordinal\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-28\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
  Matt Owen, Kosuke Imai, Gary King, and Olivia Lau

  Maintainer: Matt Owen<mowen@iq.harvard.edu>
}
\references{
}
\keyword{ package }
\keyword{ Zelig }
\keyword{ ordinal }
\keyword{ probit }
\keyword{ logit }
\seealso{
  
}
\examples{
}
