\name{zelig2ologit}
\alias{zelig2ologit}
\title{Interface between ologit model and Zelig...}
\usage{zelig2ologit(formula, weights=1, ..., data)}
\description{Interface between \code{ologit} model and Zelig}
\note{This function is exclusively for use by the \code{zelig} function}
\value{a list to be coerced into a zelig.call object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{formula}{a \code{formula}}
\item{weights}{a numeric vector}
\item{...}{ignored parameters}
\item{data}{a \code{data.frame}}}
